requirements.txt
tensorflow
streamlit
numpy
matplotlib
scikit-learn
pillow